requirements.txt
# --- Web Server & Dashboard ---
fastapi
uvicorn

# --- AI Agent Frameworks ---
google-genai
google-adk
ollama
openai
litellm
# --- Google Calendar API ---
google-auth-oauthlib
google-api-python-client

# --- Utilities ---
python-dotenv
requests
pydantic
pytz
protobuf